+++ src/components/SignIn/RegistrationLink/RegistrationLink.jsx
@@ -1,5 +1,7 @@
 import { RegistrationLinkStyled } from './RegistrationLink.styled';
 
 export default function RegistrationLink() {
-  return <RegistrationLinkStyled href="">Registration</RegistrationLinkStyled>;
+  return (
+    <RegistrationLinkStyled to="/register">Registration</RegistrationLinkStyled>
+  );
 }
