+++ src/components/SignIn/RegistrationLink/RegistrationLink.styled.js
@@ -1,6 +1,7 @@
 import { styled } from 'styled-components';
+import { Link } from 'react-router-dom';
 
-export const RegistrationLinkStyled = styled.a`
+export const RegistrationLinkStyled = styled(Link)`
   color: #fafafa;
   font-family: Poppins;
   font-size: 14px;
@@ -8,7 +9,7 @@ export const RegistrationLinkStyled = styled.a`
   font-weight: 400;
 
   text-decoration-line: underline;
-  @media screen and (min-width: 768px) {
+  @media screen and (min-width: 375px) {
     font-size: 16px;
   }
 `;
